Use ImageSurface instead of Pixbuf
