+++ src/Gtk/Avalonia.Gtk/FramebufferManager.cs
@@ -6,7 +6,7 @@ namespace Avalonia.Gtk
     class FramebufferManager : IFramebufferPlatformSurface, IDisposable
     {
         private readonly WindowImplBase _window;
-        private PixbufFramebuffer _fb;
+        private SurfaceFramebuffer _fb;
 
         public FramebufferManager(WindowImplBase window)
         {
@@ -29,8 +29,8 @@ namespace Avalonia.Gtk
             if (_fb == null || _fb.Width != width ||
                 _fb.Height != height)
             {
-                _fb?.Dispose();
-                _fb = new PixbufFramebuffer(width, height);
+                _fb?.Deallocate();
+                _fb = new SurfaceFramebuffer(width, height);
             }
             _fb.SetDrawable(drawable);
             return _fb;
